# openapi_specs/openapi.yaml
components:
  schemas:
...
    ErrorResponseModel:
      properties:
        error:
          description: Error type
          title: Error
          type: string
        error_code:
          anyOf:
          - type: string
          - type: 'null'
          description: Custom error code or identifier
          title: Error Code
        message:
          description: Human-readable error message
          title: Message
          type: string
      required:
      - error
      - message
      title: ErrorResponseModel
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    PriorityEnum:
      enum:
      - low
      - medium
      - high
      - critical
      title: PriorityEnum
      type: string
    RowCreate:
      properties:
        name:
          maxLength: 255
          minLength: 1
          title: Name
          type: string
        position:
          default: 0
          minimum: 0.0
          title: Position
          type: integer
      required:
      - name
      title: RowCreate
      type: object
    RowOwnerRead:
      properties:
        avatarUrl:
          anyOf:
          - type: string
          - type: 'null'
          title: Avatarurl
        email:
          title: Email
          type: string
        firstName:
          title: Firstname
          type: string
        id:
          format: uuid
          title: Id
          type: string
        lastName:
          title: Lastname
          type: string
      required:
      - id
      - firstName
      - lastName
      - email
      title: RowOwnerRead
      type: object
    RowRead:
      properties:
        createdAt:
          format: date-time
          title: Createdat
          type: string
        dueDate:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Duedate
        id:
          format: uuid
          title: Id
          type: string
        name:
          title: Name
          type: string
        owners:
          default: []
          items:
            $ref: '#/components/schemas/RowOwnerRead'
          title: Owners
          type: array
        position:
          title: Position
          type: integer
        priority:
          $ref: '#/components/schemas/PriorityEnum'
        status:
          $ref: '#/components/schemas/StatusEnum'
        tableId:
          format: uuid
          title: Tableid
          type: string
        updatedAt:
          format: date-time
          title: Updatedat
          type: string
      required:
      - id
      - tableId
      - name
      - status
      - priority
      - position
      - createdAt
      - updatedAt
      title: RowRead
      type: object
    RowUpdate:
      properties:
        dueDate:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Duedate
        name:
          anyOf:
          - maxLength: 255
            minLength: 1
            type: string
          - type: 'null'
          title: Name
        owners:
          anyOf:
          - items:
              format: uuid
              type: string
            maxItems: 50
            type: array
          - type: 'null'
          title: Owners
        position:
          anyOf:
          - type: integer
          - type: 'null'
          title: Position
        priority:
          anyOf:
          - $ref: '#/components/schemas/PriorityEnum'
          - type: 'null'
        status:
          anyOf:
          - $ref: '#/components/schemas/StatusEnum'
          - type: 'null'
      title: RowUpdate
      type: object
    StatusEnum:
      enum:
      - not_started
      - stuck
      - working_on_it
      - done
      title: StatusEnum
      type: string
    TableCreate:
      properties:
        color:
          anyOf:
          - type: string
          - type: 'null'
          default: '#579bfc'
          title: Color
        description:
          anyOf:
          - maxLength: 1000
            minLength: 1
            type: string
          - type: 'null'
          title: Description
        name:
          maxLength: 50
          minLength: 1
          title: Name
          type: string
        position:
          anyOf:
          - minimum: 0.0
            type: integer
          - type: 'null'
          default: 0
          title: Position
      required:
      - name
      title: TableCreate
      type: object
    TableRead:
      properties:
        boardId:
          description: Foreign key for board
          format: uuid
          title: Boardid
          type: string
        color:
          description: Table color
          title: Color
          type: string
        createdAt:
          description: Table creation date
          format: date-time
          title: Createdat
          type: string
        description:
          anyOf:
          - type: string
          - type: 'null'
          description: Table description
          title: Description
        id:
          description: Primary key for table
          format: uuid
          title: Id
          type: string
        name:
          description: Table name
          title: Name
          type: string
        position:
          description: Table order
          title: Position
          type: integer
        rows:
          description: List of rows in the table
          items:
            $ref: '#/components/schemas/RowRead'
          title: Rows
          type: array
        updatedAt:
          description: Table update date
          format: date-time
          title: Updatedat
          type: string
      required:
      - id
      - boardId
      - name
      - description
      - position
      - color
      - createdAt
      - updatedAt
      - rows
      title: TableRead
      type: object
    TableUpdate:
      properties:
        color:
          anyOf:
          - type: string
          - type: 'null'
          title: Color
        description:
          anyOf:
          - maxLength: 1000
            minLength: 1
            type: string
          - type: 'null'
          title: Description
        name:
          anyOf:
          - maxLength: 50
            minLength: 1
            type: string
          - type: 'null'
          title: Name
        position:
          anyOf:
          - minimum: 0.0
            type: integer
          - type: 'null'
          title: Position
      title: TableUpdate
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  title: Monday Lite API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/v1/boards/{board_id}/tables/:
    get:
      description: List all tables
      operationId: list_tables
      parameters:
      - in: path
        name: board_id
        required: true
        schema:
          format: uuid
          title: Board Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TableRead'
                title: Response List Tables Api V1 Boards  Board Id  Tables  Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          content:
            application/json:
              example:
                error:
                  code: INTERNAL_SERVER_ERROR
                  message: An unexpected error occurred. Please try again later.
              schema:
                $ref: '#/components/schemas/ErrorResponseModel'
          description: Internal Server Error
      summary: List Tables
      tags:
      - table
    post:
      description: Create a new table
      operationId: create_table
      parameters:
      - in: path
        name: board_id
        required: true
        schema:
          format: uuid
          title: Board Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TableCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          content:
            application/json:
              example:
                error:
                  code: INTERNAL_SERVER_ERROR
                  message: An unexpected error occurred. Please try again later.
              schema:
                $ref: '#/components/schemas/ErrorResponseModel'
          description: Internal Server Error
      summary: Create Table
      tags:
      - table
  /api/v1/boards/{board_id}/tables/{table_id}:
    delete:
      description: Delete a table by ID
      operationId: delete_table
      parameters:
      - in: path
        name: board_id
        required: true
        schema:
          format: uuid
          title: Board Id
          type: string
      - in: path
        name: table_id
        required: true
        schema:
          format: uuid
          title: Table Id
          type: string
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          content:
            application/json:
              example:
                error:
                  code: INTERNAL_SERVER_ERROR
                  message: An unexpected error occurred. Please try again later.
              schema:
                $ref: '#/components/schemas/ErrorResponseModel'
          description: Internal Server Error
      summary: Delete Table
      tags:
      - table
    patch:
      description: Update a table by ID
      operationId: update_table
      parameters:
      - in: path
        name: board_id
        required: true
        schema:
          format: uuid
          title: Board Id
          type: string
      - in: path
        name: table_id
        required: true
        schema:
          format: uuid
          title: Table Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TableUpdate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          content:
            application/json:
              example:
                error:
                  code: INTERNAL_SERVER_ERROR
                  message: An unexpected error occurred. Please try again later.
              schema:
                $ref: '#/components/schemas/ErrorResponseModel'
          description: Internal Server Error
      summary: Update Table
      tags:
      - table
  /api/v1/boards/{board_id}/tables/{table_id}/duplicate:
    post:
      description: Duplicate a table by ID
      operationId: duplicate_table
      parameters:
      - in: path
        name: board_id
        required: true
        schema:
          format: uuid
          title: Board Id
          type: string
      - in: path
        name: table_id
        required: true
        schema:
          format: uuid
          title: Table Id
          type: string
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          content:
            application/json:
              example:
                error:
                  code: INTERNAL_SERVER_ERROR
                  message: An unexpected error occurred. Please try again later.
              schema:
                $ref: '#/components/schemas/ErrorResponseModel'
          description: Internal Server Error
      summary: Duplicate Table
      tags:
      - table
  /api/v1/boards/{board_id}/tables/{table_id}/rows/:
    post:
      operationId: create_row
      parameters:
      - in: path
        name: table_id
        required: true
        schema:
          format: uuid
          title: Table Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RowCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RowRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          content:
            application/json:
              example:
                error:
                  code: INTERNAL_SERVER_ERROR
                  message: An unexpected error occurred. Please try again later.
              schema:
                $ref: '#/components/schemas/ErrorResponseModel'
          description: Internal Server Error
      summary: Create Row
      tags:
      - row
  /api/v1/boards/{board_id}/tables/{table_id}/rows/{row_id}:
    delete:
      operationId: delete_row
      parameters:
      - in: path
        name: table_id
        required: true
        schema:
          format: uuid
          title: Table Id
          type: string
      - in: path
        name: row_id
        required: true
        schema:
          format: uuid
          title: Row Id
          type: string
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          content:
            application/json:
              example:
                error:
                  code: INTERNAL_SERVER_ERROR
                  message: An unexpected error occurred. Please try again later.
              schema:
                $ref: '#/components/schemas/ErrorResponseModel'
          description: Internal Server Error
      summary: Delete Row
      tags:
      - row
    patch:
      operationId: update_row
      parameters:
      - in: path
        name: table_id
        required: true
        schema:
          format: uuid
          title: Table Id
          type: string
      - in: path
        name: row_id
        required: true
        schema:
          format: uuid
          title: Row Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RowUpdate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RowRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          content:
            application/json:
              example:
                error:
                  code: INTERNAL_SERVER_ERROR
                  message: An unexpected error occurred. Please try again later.
              schema:
                $ref: '#/components/schemas/ErrorResponseModel'
          description: Internal Server Error
      summary: Update Row
      tags:
      - row
  /api/v1/boards/{board_id}/tables/{table_id}/rows/{row_id}/duplicate:
    post:
      description: Duplicate a row
      operationId: duplicate_row
      parameters:
      - in: path
        name: row_id
        required: true
        schema:
          format: uuid
          title: Row Id
          type: string
      - in: path
        name: table_id
        required: true
        schema:
          format: uuid
          title: Table Id
          type: string
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RowRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          content:
            application/json:
              example:
                error:
                  code: INTERNAL_SERVER_ERROR
                  message: An unexpected error occurred. Please try again later.
              schema:
                $ref: '#/components/schemas/ErrorResponseModel'
          description: Internal Server Error
      summary: Duplicate Row
      tags:
      - row
  /api/v1/boards/{board_id}/tables/{table_id}/rows/{row_id}/owners/{owner_id}:
    delete:
      operationId: remove_owner
      parameters:
      - in: path
        name: table_id
        required: true
        schema:
          format: uuid
          title: Table Id
          type: string
      - in: path
        name: row_id
        required: true
        schema:
          format: uuid
          title: Row Id
          type: string
      - in: path
        name: owner_id
        required: true
        schema:
          format: uuid
          title: Owner Id
          type: string
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          content:
            application/json:
              example:
                error:
                  code: INTERNAL_SERVER_ERROR
                  message: An unexpected error occurred. Please try again later.
              schema:
                $ref: '#/components/schemas/ErrorResponseModel'
          description: Internal Server Error
      summary: Remove Owner
      tags:
      - row
    post:
      operationId: add_owner
      parameters:
      - in: path
        name: table_id
        required: true
        schema:
          format: uuid
          title: Table Id
          type: string
      - in: path
        name: row_id
        required: true
        schema:
          format: uuid
          title: Row Id
          type: string
      - in: path
        name: owner_id
        required: true
        schema:
          format: uuid
          title: Owner Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RowOwnerRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          content:
            application/json:
              example:
                error:
                  code: INTERNAL_SERVER_ERROR
                  message: An unexpected error occurred. Please try again later.
              schema:
                $ref: '#/components/schemas/ErrorResponseModel'
          description: Internal Server Error
      summary: Add Owner
      tags:
      - row
...